BACKPORT: compiler, clang: suppress warning for unused static inline …
…functions

GCC explicitly does not warn for unused static inline functions for
-Wunused-function.  The manual states:

Warn whenever a static function is declared but not defined or a
non-inline static function is unused.

Clang does warn for static inline functions that are unused.

It turns out that suppressing the warnings avoids potentially
complex #ifdef directives, which also reduces LOC.

Suppress the warning for clang.